requirements.txt
mysql-connector-python
pandas
openpyxl
python-dotenv
python-dateutil
pydantic
